added more actions buttons tests working on theme toggle tests
